xGetOdoReadings( ) added and tested for 16lakh values
1. xGetOdoReadings( ) added
2. Changed odometer variable type from uint16_t to uint32_t  to get 16 lakh odo values in all MVP files.

+++ eStar_Cluster_Demo/TouchGFX/gui/include/gui/model/Model.hpp
@@ -24,11 +24,11 @@ public:
     //void FuelData();
 protected:
     ModelListener* modelListener;
-    uint32_t counter;
+    uint16_t counter;
     //uint16_t FuelCounter;
     uint16_t TickCount;
     uint16_t speedcounter;
-    uint16_t odometer ;
+    uint32_t odometer ;
 };
 
 #endif // MODEL_HPP
