+++ eStar_Cluster_Demo/TouchGFX/gui/include/gui/model/ModelListener.hpp
@@ -16,7 +16,7 @@ public:
     }
 
     virtual void notifyCounterChanged (uint16_t counter) {};
-    virtual void notifyOdoDataChanged(uint16_t odometer){};
+    virtual void notifyOdoDataChanged(uint32_t odometer){};
     //virtual void notifyFuelCounter(int FuelCounter){};
 
 protected:
