+++ eStar_Cluster_Demo/TouchGFX/gui/include/gui/screen1_screen/Screen1Presenter.hpp
@@ -26,7 +26,7 @@ public:
     virtual void deactivate();
 
     virtual void notifyCounterChanged(uint16_t newCounter);
-    virtual void notifyOdoDataChanged(uint16_t newodoData);
+    virtual void notifyOdoDataChanged(uint32_t newodoData);
    // virtual void notifyFuelCounter(int newFuelCount);
 
     virtual ~Screen1Presenter() {}
