+++ eStar_Cluster_Demo/TouchGFX/gui/include/gui/screen1_screen/Screen1View.hpp
@@ -17,7 +17,7 @@ public:
     virtual void FuelbarWarningIcon(int newCounter);
     virtual void KMPHtoMPH(int newCounter);
     virtual void FuelGauageAnimation(int newCounter);
-    virtual void OdoDataUpdate(uint16_t newodoData);
+    virtual void OdoDataUpdate(uint32_t newodoData);
 protected:
     touchgfx::BitmapId barImageIds[10];
     uint16_t newCounter;
